split taxadb.taxadb file into 2 files accessionid.py and taxid.py for AccesionID and TaxID Python classes
