+++ doc/taxadb/accessionid.rst
@@ -1,0 +1,10 @@
+.. _accessionid:
+
+
+accessionid API reference
+=========================
+
+.. automodule:: taxadb.accessionid
+  :members:
+  :private-members:
+  :special-members:
