+++ doc/taxadb/api.rst
@@ -8,9 +8,11 @@ Contents:
 .. toctree::
    :maxdepth: 2
 
+   accessionid.rst
    app.rst
    parser.rst
    taxadb.rst
+   taxid.rst
    schema.rst
    util.rst
