+++ taxadb/taxadb.py
@@ -1,10 +1,13 @@
-from taxadb.schema import *
+from taxadb.schema import db, DatabaseFactory
 from taxadb.util import fatal
 import sys
 
 
 class TaxaDB(object):
     """Main TaxaDB package class
+
+    Parent class of the Taxadb application. Use this class to create inheriting
+    classes.
 
     Args:
         dbname (:obj:`str`): Database name to connect to
@@ -75,7 +78,7 @@ class TaxaDB(object):
         """
         if len(ids) > TaxaDB.MAX_LIST:
             fatal("Too many accession entries to request (%d), max %d" % (
-                        len(ids), AccessionID.MAX_LIST))
+                        len(ids), TaxaDB.MAX_LIST))
         return True
 
     def _unmapped_taxid(self, acc, do_exit=False):
@@ -94,238 +97,3 @@ class TaxaDB(object):
         if do_exit:
             sys.exit(1)
         return True
-
-
-class AccessionID(TaxaDB):
-    """Main accession class
-
-    Provide methods to request accession table and get associated taxonomy for
-        accession ids.
-
-    Args:
-        dbtype (:obj:`str`): Database to connect to
-        dbtype (:obj:`str`): Database type to connect to (`sqlite`, `postgre`,
-            `mysql`). Default `sqlite`
-        **kwargs: Arbitrary arguments. Supported (username, password, port,
-            hostname)
-
-    Raises:
-        SystemExit: If table `accession` does not exist
-    """
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.check_table_exists(Accession)
-
-    def taxid(self, acc_number_list):
-        """Get taxonomy of accession ids
-
-        Given a list of accession numbers, yield the accession number and their
-        associated taxids as tuples
-
-        Args:
-            acc_number_list (:obj:`list`): a list of accession numbers
-
-        Yields:
-            tuple: (accession id, taxonomy id)
-
-        """
-        self.check_list_ids(acc_number_list)
-        with db.atomic():
-            query = Accession.select().where(
-                Accession.accession << acc_number_list)
-            for i in query:
-                try:
-                    yield (i.accession, i.taxid.ncbi_taxid)
-                except Taxa.DoesNotExist:
-                    self._unmapped_taxid(i.accession)
-
-    def sci_name(self, acc_number_list):
-        """Get taxonomic scientific name for accession ids
-
-        Given a list of acession numbers, yield the accession number and their
-            associated scientific name as tuples
-
-        Args:
-            acc_number_list (:obj:`list`): a list of accession numbers
-
-        Yields:
-            tuple: (accession id, taxonomy id)
-
-        """
-        self.check_list_ids(acc_number_list)
-        with db.atomic():
-            query = Accession.select().where(
-                Accession.accession << acc_number_list)
-            for i in query:
-                try:
-                    yield (i.accession, i.taxid.tax_name)
-                except Taxa.DoesNotExist:
-                    self._unmapped_taxid(i.accession)
-
-    def lineage_id(self, acc_number_list):
-        """Get taxonomic lineage name for accession ids
-
-        Given a list of accession numbers, yield the accession number and their
-            associated lineage (in the form of taxids) as tuples
-
-        Args:
-            acc_number_list (:obj:`list`): a list of accession numbers
-
-        Yields:
-            tuple: (accession id, lineage list)
-
-        """
-        self.check_list_ids(acc_number_list)
-        with db.atomic():
-            query = Accession.select().where(
-                Accession.accession << acc_number_list)
-            for i in query:
-                try:
-                    lineage_list = []
-                    current_lineage = i.taxid.tax_name
-                    current_lineage_id = i.taxid.ncbi_taxid
-                    parent = i.taxid.parent_taxid
-                    while current_lineage != 'root':
-                        lineage_list.append(current_lineage_id)
-                        new_query = Taxa.get(Taxa.ncbi_taxid == parent)
-                        current_lineage = new_query.tax_name
-                        current_lineage_id = new_query.ncbi_taxid
-                        parent = new_query.parent_taxid
-                    yield (i.accession, lineage_list)
-                except Taxa.DoesNotExist:
-                    self._unmapped_taxid(i.accession)
-
-    def lineage_name(self, acc_number_list):
-        """Get a lineage name for accession ids
-
-        Given a list of acession numbers, yield the accession number and their
-            associated lineage as tuples
-
-        Args:
-            acc_number_list (:obj:`list`): a list of accession numbers
-
-        Yields:
-            tuple: (accession id, lineage name)
-
-        """
-        self.check_list_ids(acc_number_list)
-        with db.atomic():
-            query = Accession.select().where(
-                Accession.accession << acc_number_list)
-            for i in query:
-                try:
-                    lineage_list = []
-                    current_lineage = i.taxid.tax_name
-                    parent = i.taxid.parent_taxid
-                    while current_lineage != 'root':
-                        lineage_list.append(current_lineage)
-                        new_query = Taxa.get(Taxa.ncbi_taxid == parent)
-                        current_lineage = new_query.tax_name
-                        parent = new_query.parent_taxid
-                    yield (i.accession, lineage_list)
-                except Taxa.DoesNotExist:
-                    self._unmapped_taxid(i.accession)
-
-
-class TaxID(TaxaDB):
-    """Main class for querying taxid
-
-    Provide methods to request taxa table and get associated accession ids.
-
-    Args:
-        dbtype (:obj:`str`): Database to connect to
-        dbtype (:obj:`str`): Database type to connect to (`sqlite`, `postgre`,
-            `mysql`). Default `sqlite`
-        **kwargs: Arbitrary arguments. Supported (username, password, port,
-            hostname)
-
-    Raises:
-        SystemExit: If table `taxa` does not exist
-
-    """
-
-    def __init__(self, dbtype='sqlite', dbname=None, **kwargs):
-        super().__init__(dbtype=dbtype, dbname=dbname, **kwargs)
-        self.check_table_exists(Taxa)
-
-    def sci_name(self, taxid):
-        """Get taxonomic scientific name for taxonomy id
-
-        Given a taxid, return its associated scientific name
-
-        Args:
-            taxid (:obj:`int`): a taxid
-        Returns:
-            str: name, scientific name or None if taxid not found
-
-        """
-        try:
-            name = Taxa.get(Taxa.ncbi_taxid == taxid).tax_name
-            return name
-        except Taxa.DoesNotExist as err:
-            return None
-
-    def lineage_id(self, taxid, reverse=False):
-        """Get lineage for a taxonomic id
-
-        Given a taxid, return its associated lineage (in the form of a list of
-            taxids, each parents of each others)
-
-        Args:
-            taxid (:obj:`int`): a taxid
-            reverse (:obj:`bool`): Inverted lineage, from top to bottom
-                taxonomy hierarchy. Default False
-        Returns:
-            list: lineage_list, associated lineage id with taxid or None if
-                taxid not found
-
-        """
-        try:
-            lineage_list = []
-            current_lineage = Taxa.get(Taxa.ncbi_taxid == taxid).tax_name
-            current_lineage_id = Taxa.get(Taxa.ncbi_taxid == taxid).ncbi_taxid
-            parent = Taxa.get(Taxa.ncbi_taxid == taxid).parent_taxid
-            while current_lineage != 'root':
-                lineage_list.append(current_lineage_id)
-                new_query = Taxa.get(Taxa.ncbi_taxid == parent)
-
-                current_lineage = new_query.tax_name
-                current_lineage_id = new_query.ncbi_taxid
-                parent = new_query.parent_taxid
-            if reverse is True:
-                lineage_list.reverse()
-            return lineage_list
-        except Taxa.DoesNotExist as err:
-            return None
-
-    def lineage_name(slef, taxid, reverse=False):
-        """Get a lineage name for a taxonomic id
-
-        Given a taxid, return its associated lineage
-
-        Arguments:
-            taxid (:obj:`int`): a taxid
-            reverse (:obj:`bool`): Inverted lineage, from top to bottom
-                taxonomy hierarchy. Default False
-
-        Returns:
-            list: lineage_name, associated lineage name with taxid or None if
-                taxid not found
-
-        """
-        try:
-            lineage_list = []
-            current_lineage = Taxa.get(Taxa.ncbi_taxid == taxid).tax_name
-            parent = Taxa.get(Taxa.ncbi_taxid == taxid).parent_taxid
-            while current_lineage != 'root':
-                lineage_list.append(current_lineage)
-                new_query = Taxa.get(Taxa.ncbi_taxid == parent)
-
-                current_lineage = new_query.tax_name
-                parent = new_query.parent_taxid
-            if reverse is True:
-                lineage_list.reverse()
-            return lineage_list
-        except Taxa.DoesNotExist as err:
-            return None
